Moved navigation classes to separate module.
