+++ gradle.properties
@@ -17,11 +17,12 @@ org.gradle.jvmargs=-Xmx2048m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryErro
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
-VERSION_NAME=0.2.0
+VERSION_NAME=0.3.0-SNAPSHOT
 VERSION_CODE=4
 GROUP=io.github.azaiats.androidmvvm
 
-DESCRIPTION=A Model-View-ViewModel library for Android apps
+CORE_DESCRIPTION=A Model-View-ViewModel library for Android apps
+NAVIGATION_DESCRIPTION=Navigation support for android-mvvm library
 WEBSITE=http://a-zaiats.github.io/android-mvvm/
 ISSUE_TRACKER=https://github.com/A-Zaiats/android-mvvm/issues
 REPOSITORY=https://github.com/A-Zaiats/android-mvvm.git
