+++ navigation/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="io.github.azaiats.androidmvvm.navigation" />
