+++ sample/build.gradle
@@ -25,6 +25,7 @@ android {
 
 dependencies {
     compile project(':core')
+    compile project(':navigation')
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile "junit:junit:$rootProject.ext.junit_version"
     compile "com.android.support:appcompat-v7:$rootProject.ext.support_version"
