+++ sample/src/main/java/io/github/azaiats/androidmvvm/sample/viewmodels/MainViewModel.java
@@ -1,6 +1,6 @@
 package io.github.azaiats.androidmvvm.sample.viewmodels;
 
-import io.github.azaiats.androidmvvm.core.common.NavigatingViewModel;
+import io.github.azaiats.androidmvvm.navigation.common.NavigatingViewModel;
 import io.github.azaiats.androidmvvm.sample.navigators.MainNavigator;
 import io.github.azaiats.androidmvvm.sample.utils.AppPreferences;
