+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':sample', ':core'
+include ':sample', ':core', ':navigation', ':testutils'
