+++ testutils/build.gradle
@@ -1,0 +1,5 @@
+apply plugin: 'java'
+
+dependencies {
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+}
